refactor: extract desktop and quick nav as a single InLineNav component
