--- frontend/src/components/layout/navbar/QuickNav.tsx
@@ -1,29 +1,0 @@
-import { NavLink } from 'react-router-dom';
-
-import { type NavProps } from '@/types/navbar';
-
-const QuickNav = ({ navItems }: NavProps) => {
-  return (
-    <div className="flex items-center gap-4">
-      {navItems.map((item) => (
-        <NavLink
-          key={item.to}
-          to={item.to}
-          className={({ isActive }) =>
-            [
-              // Stryker disable next-line StringLiteral: base layout/style classes are visual-only
-              'rounded-md px-2 py-2 text-sm font-medium transition-colors',
-              isActive
-                ? 'bg-accent text-accent-foreground'
-                : 'text-muted-foreground hover:bg-accent hover:text-accent-foreground',
-            ].join(' ')
-          }
-        >
-          {item.label}
-        </NavLink>
-      ))}
-    </div>
-  );
-};
-
-export default QuickNav;
